+++ CHANGELOG.md
@@ -4,3 +4,40 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
+
+## v0.2.0
+
+[compare changes](https://github.com/Weburz/carousel/compare/v0.1.1...v0.2.0)
+
+### 🚀 Enhancements
+
+- Support breakpoint maps for slidesPerView with SSR-correct media queries ([8a8a416](https://github.com/Weburz/carousel/commit/8a8a416))
+
+### ❤️ Contributors
+
+- Sagar-Kap ([@sagar-kap](https://github.com/sagar-kap))
+
+## v0.1.1
+
+[compare changes](https://github.com/Weburz/carousel/compare/v0.1.0...v0.1.1)
+
+### 🏡 Chore
+
+- Auto hide arrows with single carousel item ([1a34e41](https://github.com/Weburz/carousel/commit/1a34e41))
+
+### ❤️ Contributors
+
+- Sagar-Kap ([@sagar-kap](https://github.com/sagar-kap))
+
+## v0.1.0
+
+
+### 🏡 Chore
+
+- Scaffold the carousels ([22fd3b9](https://github.com/Weburz/carousel/commit/22fd3b9))
+- Namespace CSS variables under --weburz- prefix ([3114039](https://github.com/Weburz/carousel/commit/3114039))
+- Add umami analytics to the playground ([b98b50d](https://github.com/Weburz/carousel/commit/b98b50d))
+
+### ❤️ Contributors
+
+- Sagar-Kap ([@sagar-kap](https://github.com/sagar-kap))